+++ LakersProject/src/data/JDBC.java
@@ -1,4 +1,4 @@
-package guicomponents;
+package data;
 import java.sql.*;
 
 public class JDBC {
